Move test to separate class
